Fixed a bug that causes SEGFAULTs when the number of rotations of the pulsar durind the observation goes above 2G.
git-svn-id: file://localhost/data1/git/presto_svn/trunk@1225 5fac236c-a0d8-0310-9745-ef116914eda1

+++ src/prepfold.c
@@ -1092,7 +1092,7 @@ int main(int argc, char *argv[])
          }
          partnum = (int) floor(event * dtmp);
          phase = event * (event * (event * tfdd + tfd) + tf);
-         binnum = (int) ((phase - (int) phase) * search.proflen);
+         binnum = (int) ((phase - (long long) phase) * search.proflen);
          search.rawfolds[partnum * search.proflen + binnum] += 1.0;
       }
       if (binary) {
